fix query error reporting
