+++ kogito-codegen-modules/kogito-codegen-rules/src/main/java/org/kie/kogito/codegen/rules/AbstractQueryEntrypointGenerator.java
@@ -62,57 +62,4 @@ public abstract class AbstractQueryEntrypointGenerator implements RuleFileGenera
         return generator.generatedFilePath();
     }
 
-    @Override
-    public boolean validate() {
-        return !query.model().getBindings().isEmpty();
-    }
-
-    @Override
-    public DroolsError getError() {
-        if (query.model().getBindings().isEmpty()) {
-            return new NoBindingQuery(query.model());
-        }
-        return null;
-    }
-
-    public static class NoBindingQuery extends DroolsError {
-
-        private static final int[] ERROR_LINES = new int[0];
-
-        private final QueryModel query;
-
-        public NoBindingQuery(QueryModel query) {
-            this.query = query;
-        }
-
-        @Override
-        public String getMessage() {
-            return "Query " + query.getName() + " has no bound variable. At least one binding is required to determine the value returned by this query";
-        }
-
-        @Override
-        public int[] getLines() {
-            return ERROR_LINES;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-            if (!super.equals(o)) {
-                return false;
-            }
-            NoBindingQuery that = (NoBindingQuery) o;
-            return Objects.equals(query, that.query);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(super.hashCode(), query);
-        }
-    }
 }
